update packager.Pull with new api, make destination a required field, move packager2 impl back to packager2 pending full upgrade
