+++ src/cmd/package.go
@@ -1075,6 +1075,7 @@ func newPackagePullCommand(v *viper.Viper) *cobra.Command {
 }
 
 func (o *packagePullOptions) run(cmd *cobra.Command, args []string) error {
+	srcURL := args[0]
 	outputDir := pkgConfig.PullOpts.OutputDirectory
 	if outputDir == "" {
 		wd, err := os.Getwd()
@@ -1083,8 +1084,7 @@ func (o *packagePullOptions) run(cmd *cobra.Command, args []string) error {
 		}
 		outputDir = wd
 	}
-	err := packager.Pull(cmd.Context(), args[0], packager.PullOptions{
-		OutputDirectory:         outputDir,
+	err := packager.Pull(cmd.Context(), srcURL, outputDir, packager.PullOptions{
 		SHASum:                  pkgConfig.PkgOpts.Shasum,
 		SkipSignatureValidation: pkgConfig.PkgOpts.SkipSignatureValidation,
 		PublicKeyPath:           pkgConfig.PkgOpts.PublicKeyPath,
